requirements/base.txt
Jinja2==2.9.6
python-slugify==1.2.4